refactor(types): remove all functions in type repository except get all
