+++ api/handlers/record_handler.go
@@ -2,7 +2,6 @@ package handlers
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -47,20 +46,16 @@ func (h *RecordHandler) Delete(w http.ResponseWriter, r *http.Request) {
 		errMessage = fmt.Sprintf("error deleting record \"%s\" with type \"%s\"", recordID, typeName)
 	)
 
-	tName, err := h.typeRepository.GetByName(r.Context(), typeName)
-	if errors.As(err, new(record.ErrNoSuchType)) {
-		writeJSONError(w, http.StatusNotFound, bodyParserErrorMsg(err))
-		return
-	}
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, bodyParserErrorMsg(err))
+	typName := record.TypeName(typeName)
+	if err := typName.IsValid(); err != nil {
+		writeJSONError(w, http.StatusNotFound, err.Error())
 		return
 	}
 
-	err = h.discoveryService.DeleteRecord(r.Context(), tName.String(), recordID)
+	err := h.discoveryService.DeleteRecord(r.Context(), typName.String(), recordID)
 	if err != nil {
 		h.logger.
-			Errorf("error deleting record \"%s\": %v", tName, err)
+			Errorf("error deleting record \"%s\": %v", typName, err)
 
 		if _, ok := err.(record.ErrNoSuchRecord); ok {
 			statusCode = http.StatusNotFound
@@ -71,7 +66,7 @@ func (h *RecordHandler) Delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	h.logger.Infof("deleted record \"%s\" with type \"%s\"", recordID, tName)
+	h.logger.Infof("deleted record \"%s\" with type \"%s\"", recordID, typName)
 	writeJSON(w, http.StatusNoContent, "success")
 }
 
@@ -85,20 +80,16 @@ func (h *RecordHandler) UpsertBulk(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	tName, err := h.typeRepository.GetByName(r.Context(), typeName)
-	if errors.As(err, new(record.ErrNoSuchType)) {
+	typName := record.TypeName(typeName)
+	if err := typName.IsValid(); err != nil {
 		writeJSONError(w, http.StatusNotFound, err.Error())
-		return
-	}
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, bodyParserErrorMsg(err))
 		return
 	}
 
 	var failedRecords = make(map[int]string)
 	for idx, record := range records {
 		if err := h.validateRecord(record); err != nil {
-			h.logger.WithField("type", tName).
+			h.logger.WithField("type", typName).
 				WithField("record", record).
 				Errorf("error validating record: %v", err)
 			failedRecords[idx] = err.Error()
@@ -109,34 +100,30 @@ func (h *RecordHandler) UpsertBulk(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := h.discoveryService.Upsert(r.Context(), tName.String(), records); err != nil {
-		h.logger.WithField("type", tName).
+	if err := h.discoveryService.Upsert(r.Context(), typName.String(), records); err != nil {
+		h.logger.WithField("type", typName).
 			Errorf("error creating/updating records: %v", err)
 
 		status := http.StatusInternalServerError
 		writeJSONError(w, status, http.StatusText(status))
 		return
 	}
-	h.logger.Infof("created/updated %d records for %q type", len(records), tName)
+	h.logger.Infof("created/updated %d records for %q type", len(records), typName)
 	writeJSON(w, http.StatusOK, StatusResponse{Status: "success"})
 }
 
 func (h *RecordHandler) GetByType(w http.ResponseWriter, r *http.Request) {
 	typeName := mux.Vars(r)["name"]
 
-	tName, err := h.typeRepository.GetByName(r.Context(), typeName)
-	if errors.As(err, new(record.ErrNoSuchType)) {
+	typName := record.TypeName(typeName)
+	if err := typName.IsValid(); err != nil {
 		writeJSONError(w, http.StatusNotFound, err.Error())
 		return
 	}
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, bodyParserErrorMsg(err))
-		return
-	}
 
-	recordRepo, err := h.recordRepositoryFactory.For(tName.String())
+	recordRepo, err := h.recordRepositoryFactory.For(typName.String())
 	if err != nil {
-		h.logger.WithField("type", tName).
+		h.logger.WithField("type", typName).
 			Errorf("error constructing record repository: %v", err)
 		status, message := h.responseStatusForError(err)
 		writeJSONError(w, status, message)
@@ -150,7 +137,7 @@ func (h *RecordHandler) GetByType(w http.ResponseWriter, r *http.Request) {
 
 	recordList, err := recordRepo.GetAll(r.Context(), getCfg)
 	if err != nil {
-		h.logger.WithField("type", tName).
+		h.logger.WithField("type", typName).
 			Errorf("error fetching records: GetAll: %v", err)
 		status, message := h.responseStatusForError(err)
 		writeJSONError(w, status, message)
@@ -171,19 +158,15 @@ func (h *RecordHandler) GetOneByType(w http.ResponseWriter, r *http.Request) {
 		recordID = vars["id"]
 	)
 
-	tName, err := h.typeRepository.GetByName(r.Context(), typeName)
-	if errors.As(err, new(record.ErrNoSuchType)) {
-		writeJSONError(w, http.StatusNotFound, bodyParserErrorMsg(err))
-		return
-	}
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, bodyParserErrorMsg(err))
+	typName := record.TypeName(typeName)
+	if err := typName.IsValid(); err != nil {
+		writeJSONError(w, http.StatusNotFound, err.Error())
 		return
 	}
 
-	recordRepo, err := h.recordRepositoryFactory.For(tName.String())
+	recordRepo, err := h.recordRepositoryFactory.For(typName.String())
 	if err != nil {
-		h.logger.WithField("type", tName).
+		h.logger.WithField("type", typName).
 			Errorf("internal: error construing record repository: %v", err)
 
 		status := http.StatusInternalServerError
@@ -193,7 +176,7 @@ func (h *RecordHandler) GetOneByType(w http.ResponseWriter, r *http.Request) {
 
 	record, err := recordRepo.GetByID(r.Context(), recordID)
 	if err != nil {
-		h.logger.WithField("type", tName).
+		h.logger.WithField("type", typName).
 			WithField("record", recordID).
 			Errorf("error fetching record: %v", err)
