+++ store/elasticsearch/record_repository.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"net/url"
 	"strings"
+	"time"
 
 	"github.com/elastic/go-elasticsearch/v7"
 	"github.com/elastic/go-elasticsearch/v7/esapi"
@@ -20,6 +21,9 @@ import (
 const (
 	defaultGetSize   = 20
 	defaultSortField = "name.keyword"
+
+	defaultScrollTimeout   = 30 * time.Second
+	defaultScrollBatchSize = 20
 )
 
 // RecordRepository implements discovery.RecordRepository
